Refactor log history management: Introduce DailyFileLogHistory implementation
- Moved file log history management to a new class DailyFileLogHistory for optimized daily log handling.
- Updated file structure to include a dedicated directory for file log history.
- Enhanced session cleanup strategies with options for deleting oldest files, deleting by size, and archiving.
- Added comprehensive session statistics tracking for better log management insights.
- Implemented robust error handling and logging for file operations.
- Updated exports and imports to handle JSON data efficiently with chunked processing.
