+++ packages/ispectify/lib/ispectify.dart
@@ -2,7 +2,7 @@ export 'package:ansicolor/ansicolor.dart';
 
 export 'src/enums/log_type.dart';
 export 'src/filter/filter.dart';
-export 'src/history/file_log_history.dart';
+export 'src/history/file_log/file_log_history.dart';
 export 'src/history/history.dart';
 export 'src/history/serialization.dart';
 export 'src/ispectify.dart';
